# BF16 optimizer sync parameters with module when load_module_only=True