@AIproj 2023-11-26
Recommend merging, it fixes bugs similar to those mentioned [in this issue](https://github.com/microsoft/DeepSpeed/pull/4434) by implementing some of those solutions (alternatively, DeeperSpeed may need to be brought to the most recent DeepSpeed main branch).